refactor: rename EmotionThemeProvider to InstUISettingsProvider occurences
Also add deprecation guide
