+++ docs/getting-started/usage.md
@@ -23,7 +23,7 @@ This will generate a vanilla React app with tests, sample code and scripts. You 
   ...
   "dependencies": {
     ...
-    "@instructure/ui": "^8.6.9"
+    "@instructure/ui": "^8"
   }
 }
 ```
@@ -33,13 +33,13 @@ Run `yarn install`, so InstUI is downloaded to your `node_modules` folder and ca
 3. Now you are ready to use InstUI, let's try it out. Replace the code in `App.js` with the following:
 
 ```javascript
-import { Button, EmotionThemeProvider, canvas } from '@instructure/ui'
+import { Button, InstUISettingsProvider, canvas } from '@instructure/ui'
 
 function App() {
   return (
-    <EmotionThemeProvider theme={canvas}>
+    <InstUISettingsProvider theme={canvas}>
       <Button>Hello from InstUI!</Button>
-    </EmotionThemeProvider>
+    </InstUISettingsProvider>
   )
 }
 
@@ -48,7 +48,7 @@ export default App
 
 What does this code do?
 
-- [EmotionThemeProvider](#EmotionThemeProvider) adds a theme to your application. InstUI components require a theme to work, all components are themeable, and themes control their look and feel. There are 3 built-in themes: [`canvas`](#canvas), [`canvasHighContrast`](#canvas-high-contrast), [`instructure`](#instructure). The component examples seen throughout the documentation use the [canvas theme](#canvas) by default.
+- [InstUISettingsProvider](#InstUISettingsProvider) adds a theme to your application. InstUI components require a theme to work, all components are themeable, and themes control their look and feel. There are 3 built-in themes: [`canvas`](#canvas), [`canvasHighContrast`](#canvas-high-contrast), [`instructure`](#instructure). The component examples seen throughout the documentation use the [canvas theme](#canvas) by default.
 - [Button](#Button) is an Instructure UI button component
 
 Finally, run `yarn start` to start up a basic development server.
@@ -57,10 +57,10 @@ Congrats, you have now a (very) basic app that uses Instructure UI :)
 
 ## Integrating With an Existing Project
 
-Just add the `@instructure/ui` dependency as shown above and wrap the part of your app that will use InstUI in `<EmotionThemeProvider>` and start using InstUI components.
+Just add the `@instructure/ui` dependency as shown above and wrap the part of your app that will use InstUI in `<InstUISettingsProvider>` and start using InstUI components.
 
 ## Further reading
 
-- To use a different theme or customize one read about [EmotionThemeProvider](#EmotionThemeProvider)
+- To use a different theme or customize one read about [InstUISettingsProvider](#InstUISettingsProvider)
 - Make sure you read about [Accessibility](#accessibility) with InstUI.
 - [How to make your own component that uses InstUI's theming engine](#emotion)
