+++ packages/__docs__/components.js
@@ -121,5 +121,8 @@ export { Guidelines } from './src/Guidelines'
 export { Figure } from './src/Figure'
 // eslint-disable-next-line no-restricted-imports
 export { ToggleBlockquote } from './src/ToggleBlockquote'
-export { EmotionThemeProvider } from '@instructure/emotion'
+export {
+  EmotionThemeProvider,
+  InstUISettingsProvider
+} from '@instructure/emotion'
 export { canvas, canvasHighContrast, instructure } from '@instructure/ui-themes'
