+++ packages/emotion/src/EmotionThemeProvider/README.md
@@ -2,6 +2,10 @@
 describes: EmotionThemeProvider
 ---
 
+#### DEPRECATED. Please use [InstUISettingsProvider](#InstUISettingsProvider) instead. IT has the same functionality and
+
+adds text direction configuration.
+
 The `<EmotionThemeProvider/>` component provides a way to set the theme for our app or override the default theme properties for all themeable child components using the [withStyle](#withStyle) decorator.
 
 Note that `<EmotionThemeProvider/>` components can be nested and that
