+++ packages/template-app/template/README.md.ejs
@@ -50,7 +50,7 @@ which is also the default for our documentation.
 
 Alternatively, if you want to continue using the `canvas` theme as your default
 with some overrides instead of creating an entirely new one, you can use the
-[EmotionThemeProvider](https://instructure.design/#EmotionThemeProvider) component.
+[InstUISettingsProvider](https://instructure.design/#InstUISettingsProvider) component.
 
 You can also make your custom built components in this app themeable. The `Banner`
 component located in the source directory, for example, is already using
