+++ packages/template-app/template/src/index.js.ejs
@@ -26,10 +26,10 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
 import { canvas } from '@instructure/ui-themes'
-import { EmotionThemeProvider } from '@instructure/emotion'
+import { InstUISettingsProvider } from '@instructure/emotion'
 
 ReactDOM.render(
-  <EmotionThemeProvider theme={canvas}>
+  <InstUISettingsProvider theme={canvas}>
     <App />
-  </EmotionThemeProvider>,
+  </InstUISettingsProvider>,
   document.getElementById('app'))
